Move database model to database package.
