+++ app/src/main/java/com/example/weatherapp11/Repository.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.example.weatherapp11.apis.JsonWeatherApi
 import com.example.weatherapp11.database.WeatherDao
+import com.example.weatherapp11.database.WeatherEntity
 import com.example.weatherapp11.models.WeatherInfo
 import retrofit2.Call
 import retrofit2.Callback
