+++ app/src/main/java/com/example/weatherapp11/database/WeatherDao.kt
@@ -1,7 +1,6 @@
 package com.example.weatherapp11.database
 
 import androidx.room.*
-import com.example.weatherapp11.WeatherEntity
 
 @Dao
 interface WeatherDao {
