+++ app/src/main/java/com/example/weatherapp11/database/WeatherDatabase.kt
@@ -5,7 +5,6 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.sqlite.db.SupportSQLiteDatabase
-import com.example.weatherapp11.WeatherEntity
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
